backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic==2.9.2
pytz==2024.1
requests==2.32.3
numpy==1.26.4
opencv-python==4.10.0.84
ultralytics==8.3.28
Pillow==10.4.0
lapx==0.9.2
psutil==6.0.0
python-dotenv==1.0.1
anyio==4.4.0
starlette==0.38.5
